Task 11

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-unsigned n, i, max;
+unsigned long n, i, max;
 
 n = 612852475143;
 max = 1;
